finalisation du webserviceUser
